Knowing if the user was dirty alone wasn't enough. After requesting the user, it would be marked as dirty and the batch task would update it instantly. Added a new field to know WHEN it was last updated, that way we give enough time for a test to run with that same user - this should be a configurable option anyways, since the amount of time is probably very variable.
Refactored the way queries to the database worked. We now no longer support
GET, instead a POST with a a JSON payload is more adequate for complex
requests. Also improved alot the usage of the PeeWee ORM by using reduce
which helps us dynamically setup WHERE clauses.

Added initial support for a very basic page. The goal is to have a page
where any user (instead of automation) may request data from the database
so that manual testers can also enjoy the service.